software/sorter/backend/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boxmot>=18.0.0",
    "cryptography>=42",
    "cv2-enumerate-cameras>=1.3.3",
    "fastapi>=0.128.3",
    "flask>=3.1.3",
    "lap>=0.5.12",
    "litellm>=1.83.0",
    "ncnn>=1.0.20260114",
    "numpy>=2.4.2",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.0",
    "pydantic-to-typescript>=2.0.0",
    "pyright>=1.1.408",
    "pyserial>=3.5",
    "python-dotenv>=1.2.1",
    "readchar>=4.2.1",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "supervision>=0.27.0.post2",
    "trackers>=2.3.0",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
    "wsproto>=1.3.2",
    "zeroconf>=0.148.0",
]

[tool.pyright]
reportMissingImports = "none"
reportMissingModuleSource = "none"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
